Use a dedicated executor for stdin
